pyproject.toml
[tool.poetry]
name = "project-red"
version = "0.1.0"
description = "making tast task"
authors = ["aisner"]
license = "MIT"
readme = "README.md"
packages = [{include = "project_red"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"
djangorestframework = "^3.14.0"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
